Fix fake clientsets in metrics.k8s.io.
The generated fake clientsets were using the API group "metrics" instead
of "metrics.k8s.io". This patch includes a test that fails without the
fix.

The `+groupName` annotation needed to be copied to the v1alpha1 and
v1beta1 packages to fix it.

As a result, tests using this fake clientset work as expected.

Kubernetes-commit: ab7e2ff136b8571dd1a0ac89831f2406d9f84d4f